requirements.txt
functions-framework>=3.*
aiohttp
google-cloud-storage